refactor: rename Android application ID, annotate use cases for injection
