+++ android/app/src/main/AndroidManifest.xml
@@ -1,5 +1,5 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="dev.adryanev.template">
+    package="dev.mfrozi.expense_tracker">
    <application
         android:label="${appName}"
         android:icon="@mipmap/ic_launcher">
