+++ android/app/src/main/kotlin/dev/mfrozi/expense_tracker/MainActivity.kt
@@ -1,4 +1,4 @@
-package dev.adryanev.template
+package dev.mfrozi.expense_tracker
 
 import io.flutter.embedding.android.FlutterActivity
