+++ lib/app/router/app_router.dart
@@ -2,7 +2,7 @@ import 'package:equatable/equatable.dart';
 import 'package:flutter/foundation.dart';
 import 'package:go_router/go_router.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:template/core/di/injector.dart';
+import 'package:template/injector.dart';
 import 'package:template/features/category/presentation/blocs/category_cubit.dart';
 import 'package:template/features/category/presentation/pages/category_manage_page.dart';
